refactor(web): move auth context and route guard into lib/
