+++ web/src/routes/LoginPage.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom'
-import { useAuth } from '@/features/auth/AuthContext'
+import { useAuth } from '@/lib/AuthContext'
 import { LoginForm } from '@/features/auth/LoginForm'
 
 export function LoginPage() {
